fixed some errors and moved dino wsdl tests to a notest module, maybe we can delete this tests later :(
