+++ tests/test_005_io_dino.py
@@ -8,32 +8,6 @@ from hydropandas import observation as obs
 from hydropandas.io import io_dino
 
 
-# TEST_DIR = os.path.dirname(os.path.abspath(__file__))
-# PROJECT_DIR = os.path.abspath(os.path.join(TEST_DIR, os.pardir))
-# sys.path.insert(0, PROJECT_DIR)
-# os.chdir(TEST_DIR)
-
-
-def test_dino_download_single():
-    # download single file
-    measurements, meta = io_dino.download_dino_groundwater(location="B57F0077",
-                                                           filternr=4.,
-                                                           tmin="2000-01-01",
-                                                           tmax="2010-01-01",
-                                                           unit="NAP")
-    return measurements, meta
-
-
-def test_dino_download_single_empty():
-    # download empty
-    measurements, meta = io_dino.download_dino_groundwater(location="B50E0092",
-                                                           filternr=1.,
-                                                           tmin="1900-01-01",
-                                                           tmax="2040-01-01",
-                                                           unit="NAP",
-                                                           verbose=True)
-    return measurements, meta
-
 
 def test_dino_csv():
     fname = ('./tests/data/2019-Dino-test/Grondwaterstanden_Put/'
@@ -73,34 +47,6 @@ def test_dino_metadata3():
     return meta
 
 
-def test_dino_meetreeks():
-
-    dino = io_dino.DinoWSDL()
-
-    # measurements
-    measurements = dino.findMeetreeks(location='B58A0092',
-                                      filternr='004',
-                                      tmin="1900-01-01",
-                                      tmax="2040-01-01")
-
-    return measurements
-
-
-def test_dino_download_extent():
-    # download extent
-    extent = [117850, 117980, 439550, 439700]  # Schoonhoven zoomed
-
-    import time
-    start_time = time.time()
-    gw_col = io_dino.download_dino_within_extent(extent,
-                                                 ObsClass=obs.GroundwaterObs,
-                                                 layer='grondwatermonitoring',
-                                                 verbose=True)
-
-    print("--- %s seconds ---" % (time.time() - start_time))
-    return gw_col
-
-
 def test_get_dino_locations():
     bbox = [130988.58668351, 386465.80657847, 135922.01739107, 391922.4284911]
 
@@ -108,15 +54,3 @@ def test_get_dino_locations():
                                      layer='grondwatermonitoring')
 
     return gdf
-
-
-def test_dino_download_single_empty():
-
-    measurements, meta = io_dino.download_dino_groundwater(
-        location="B50E0130",
-        filternr=1.,
-        tmin="1900-01-01",
-        tmax="2040-01-01",
-        unit="NAP",
-        verbose=True)
-    return
