+++ tests/tests/data/2019-Pystore-test/test_pystore/metadata.json
@@ -1,0 +1,1 @@
+{"engine": "fastparquet", "_updated": "2021-02-03 12:12:14.311355"}
